requirements.txt
alpaca-py
pandas
numpy
python-dotenv
streamlit
plotly
pytz
requests
streamlit-autorefresh
